Move the temp fix into a contained script file + update .gitignore comment
