+++ .gitignore
@@ -59,6 +59,6 @@ temp/
 
 .flake8
 
-# pending: Sphinx 8.1.4 + deps are verified as workin with arcade
-# this works around an annoying static copy issue
+# pending: Sphinx 8.1.4 + deps are verified as working with Arcade
+# see util/sphinx_static_file_temp_fix.py
 .ENABLE_DEVMACHINE_SPHINX_STATIC_FIX
