+++ doc/conf.py
@@ -32,59 +32,16 @@ HERE = Path(__file__).resolve()
 DOC_DIR = HERE.parent
 REPO_LOCAL_ROOT = DOC_DIR.parent
 
-# pending: Sphinx 8.1.4 + deps verified as compatible with Arcade to fix static copy
-# Make it move because Sphinx only fixed the copy issue as of a few days ago
-# https://github.com/sphinx-doc/sphinx/pull/13236
-# https://github.com/sphinx-doc/sphinx/issues/181
-ENABLE_DEVMACHINE_SPHINX_STATIC_FIX = REPO_LOCAL_ROOT / ".ENABLE_DEVMACHINE_SPHINX_STATIC_FIX"
-STATIC_SOURCE_DIR = DOC_DIR / "_static"
-
 ARCADE_MODULE = REPO_LOCAL_ROOT / "arcade"
 UTIL_DIR = REPO_LOCAL_ROOT / "util"
 BUILD_DIR = REPO_LOCAL_ROOT / "build"
 BUILD_HTML_DIR = BUILD_DIR / "html"
-BUILD_STATIC_DIR = BUILD_HTML_DIR / "_static"
+
 
 log.info(f"Absolute path for our conf.py       : {str(HERE)!r}")
 log.info(f"Absolute path for the repo root     : {str(REPO_LOCAL_ROOT)!r}")
 log.info(f"Absolute path for the arcade module : {str(REPO_LOCAL_ROOT)!r}")
 log.info(f"Absolute path for the util dir      : {str(UTIL_DIR)!r}")
-
-
-STATIC_CSS_DIR = STATIC_SOURCE_DIR / "css"
-force_copy_on_change = {  # pending: sphinx >= 8.1.4
-    source_file: BUILD_STATIC_DIR / f"css/{source_file.name}"
-    for source_file in STATIC_CSS_DIR.glob("*.css")
-}
-
-def force_sync(src, dest, dry: bool = False):
-    if sphinx.__version__ >= '8.1.4':
-        log.warning(
-            'Sphinx >= 8.1.4 may patch broken _static copy\n'
-            '  (see https://github.com/sphinx-doc/sphinx/issues/1810)')
-    try:
-        if src.read_text() != dest.read_text():
-            if dry:
-                log.info(f" DRY : {src} was out of date, but dry run left it as-is!")
-            # shutil.copyfile(src, dest)
-            else:
-                log.info(f" SYNC: {src} was out of date!")
-
-        else:
-            log.info(f" SKIP: {src} is current!")
-    except Exception as e:
-        log.error(f" FAIL: {src} failed: {e}")
-        raise e
-
-
-if not ENABLE_DEVMACHINE_SPHINX_STATIC_FIX.exists():
-    log.info(f"SKIP: Force-sync found no {ENABLE_DEVMACHINE_SPHINX_STATIC_FIX} file!")
-elif BUILD_HTML_DIR.exists():
-    log.info(f"SYNC: Force-sync enable file found")
-    for src, dest in force_copy_on_change.items():
-        force_sync(src, dest)
-else:
-    log.info("Skipping force-sync due to no build dir")
 
 
 # _temp_version = (REPO_LOCAL_ROOT / "arcade" / "VERSION").read_text().replace("-",'')
@@ -140,13 +97,14 @@ def run_util(filename, run_name="__main__", init_globals=None):
 
     runpy.run_path(full_str, **kwargs)
 
+run_util("sphinx_static_file_temp_fix.py")
+
 # Make thumbnails for the example code screenshots
 run_util("generate_example_thumbnails.py")
 # Create a tabular representation of the resources with embeds
 run_util("create_resources_listing.py", init_globals=RESOURCE_GLOBALS)
 # Run the generate quick API index script
 run_util('../util/update_quick_index.py')
-
 
 autodoc_inherit_docstrings = False
 autodoc_default_options = {
